src/microservices/dependencia/requirements.txt
Django==1.11.10
djangorestframework==3.7.7
graphene-django==2.0.0
mongoengine==0.15.0
PyJWT==1.5.3
pymongo==3.6.0
pytz==2017.3
six==1.11.0
